kxyz/requirements.txt
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-dysmsapi==1.0.0
aliyunsdkcore==1.0.3
certifi==2019.11.28
chardet==3.0.4
Django==2.2.8
django-debug-toolbar==2.1
django-shortuuidfield==0.1.3
djangorestframework==3.11.0
idna==2.8
jmespath==0.9.4
Pillow==6.2.1
pycrypto==2.6.1
pycryptodome==3.9.4
PyMySQL==0.9.3
python3-memcached==1.51
pytz==2019.3
qiniu==7.2.6
requests==2.22.0
shortuuid==0.5.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
